ESQL: Move `Failures` into the esql proper
This moves the `Failueres` and `Failure` class into ESQL's main package,
further slimming down our custom fork of the shared ql code. Slowly
slowly slowly, it will be no more.
