+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/VerificationException.java
@@ -7,8 +7,8 @@
 
 package org.elasticsearch.xpack.esql;
 
-import org.elasticsearch.xpack.esql.core.common.Failure;
-import org.elasticsearch.xpack.esql.core.common.Failures;
+import org.elasticsearch.xpack.esql.common.Failure;
+import org.elasticsearch.xpack.esql.common.Failures;
 
 import java.util.Collection;
