+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/common/Failure.java
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.esql.core.common;
+package org.elasticsearch.xpack.esql.common;
 
 import org.elasticsearch.xpack.esql.core.tree.Location;
 import org.elasticsearch.xpack.esql.core.tree.Node;
